+++ homeworks/team-4/flask-blog/flask_blog/templates/view-post.html
@@ -45,7 +45,7 @@
         <ul class="list-comments">
               {% for comment in post.comments.all() %}
             <li>
-                <p class="name">{{comment.name}}</p>
+                <p class="name">{{comment.title}}</p>
                 <p class="date-post">Posted at {{comment.time}}</p>
                 <p class="comment-text">{{ comment.text }}</p>
                 <a class="date-article" href = "{{url_for('delete_comment', id = post.id, comment_id = comment.id)}}" >Delete</a>
