+++ homeworks/team-4/flask-blog/run.py
@@ -1,0 +1,4 @@
+from flask_blog import app
+
+if __name__ =="__main__":
+    app.run()
